refactor(example-scripts): placed scripts in folder, updated docs
