+++ examples/scripts/.env.example
