+++ examples/scripts/appConfig.ts
